# solution/mainServer/swaggerDef.yaml
openapi: 3.0.0
info:
  title: Football Manager API
  version: 1.0.0
  description: This is a simple CRUD API application made with Express and documented with Swagger
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8080/api
components:
  schemas:
    Appearence:
      $ref: '/models/swaggerSchemas/appearence.yaml#/components/schemas/Appearence'
    Game:
        $ref: '/models/swaggerSchemas/game.yaml#/components/schemas/Game'
    GameEvent:
        $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/schemas/GameEvent'
    GameLineup:
        $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/schemas/GameLineup'
    Player:
        $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'
    PlayerValuation:
        $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/schemas/PlayerValuation'
    ChatRoom:
        $ref: '/models/swaggerSchemas/room.yaml#/components/schemas/ChatRoom'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /appearence:
    get:
      tags:
        - Appearence
      description: Use to request all appearences
      parameters:
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/page'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/sort'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/limit'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/fields'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/game_id'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/player_id'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/player_name'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/player_club_id'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/player_current_club_id'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/date'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/competition_id'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/yellow_cards'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/red_cards'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/goals'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/assists'
        - $ref: '/models/swaggerSchemas/appearence.yaml#/parameters/minutes_played'

      responses:
        200:
          description: A successful response
          content:
              application/json:
                schema:
                    type: array
                    items:
                      $ref: '/models/swaggerSchemas/appearence.yaml#/components/schemas/Appearence'
        404:
          description: No appearences found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "fail"
                  message:
                    type: string
                    example: "No appearences found"
    post:
      tags:
        - Appearence
      description: Use to create a new appearence
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/appearence.yaml#/components/schemas/Appearence'

      responses:
        200:
          description: A successful response
...
  /game:
    get:
      tags:
        - Game
      description: Use to request all games
      parameters:
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/page'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/sort'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/limit'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/fields'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/game_id'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/home_club_id'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/away_club_id'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/home_club_goals'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/away_club_goals'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/home_club_position'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/away_club_position'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/stadium'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/attendance'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/referee'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/url'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/home_club_formation'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/away_club_formation'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/home_club_name'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/away_club_name'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/aggregate'
        - $ref: '/models/swaggerSchemas/game.yaml#/parameters/competition_type'

      responses:
        200:
          description: A successful response
    post:
      tags:
        - Game
      description: Use to create a new game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/game.yaml#/components/schemas/Game'

      responses:
        200:
          description: A successful response
...
  /gameEvent:
    get:
      #same as above
      tags:
         - GameEvent
      description: Use to request all game events
      parameters:
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/page'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/sort'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/limit'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/fields'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/date'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/game_id'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/minute'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/type'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/club_id'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/player_id'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/description'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/player_in_id'
        - $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/parameters/player_assist_id'
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/schemas/GameEvent'
    post:
        tags:
            - GameEvent
        description: Use to create a new game event
        requestBody:
          content:
            application/json:
                schema:
                  $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/schemas/GameEvent'

        responses:
            200:
              description: A successful response
              content:
                application/json:
                    schema:
                        $ref: '/models/swaggerSchemas/gameEvent.yaml#/components/schemas/GameEvent'
...
  /gameLineup:
    get:
      tags:
        - GameLineup
      description: Use to request all game lineups
      parameters:
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/page'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/sort'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/limit'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/fields'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/game_id'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/club_id'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/type'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/number'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/player_id'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/player_name'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/team_captain'
        - $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/parameters/position'
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/schemas/GameLineup'
    post:
      tags:
        - GameLineup
      description: Use to create a new game lineup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/schemas/GameLineup'

      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '/models/swaggerSchemas/gameLineup.yaml#/components/schemas/GameLineup'
...
  /player:
    get:
      tags:
        - Player
      description: Use to request all players
      parameters:
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/page'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/sort'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/limit'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/fields'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/first_name'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/last_name'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/last_season'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/player_code'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/country_of_birth'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/city_of_birth'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/country_of_citizenship'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/date_of_birth'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/sub_position'
        - $ref: '/models/swaggerSchemas/player.yaml#/components/parameters/position'
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'
    post:
      tags:
        - Player
      description: Use to create a new player
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'

      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'
  /player/{id}:
    get:
      tags:
        - Player
      description: Use to request one player
      parameters:
        - name: id
          in: path
          description: Player ID
          required: true
          example: 6574baf3975bb4cdf6a05aee
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "6574baf3975bb4cdf6a05aee"
                      first_name:
                        type: string
                        example: "Vladyslav"
                      last_name:
                        type: string
                        example: "Supryaga"
                      last_season:
                        type: string
                        example: "2023"
                      player_code:
                        type: string
                        example: "vladyslav-supryaga"
                      country_of_birth:
                        type: string
                        example: "Ukraine"
                      city_of_birth:
                        type: string
                        example: "Sarata, Odesa Oblast"
                      country_of_citizenship:
                        type: string
                        example: "Ukraine"
                      date_of_birth:
                        type: string
                        example: "2000-02-15"
                      sub_position:
                        type: string
                        example: "Centre-Forward"
                      position:
                        type: string
                        example: "Attack"
                      foot:
                        type: string
                        example: "right"
                      height_in_cm:
                        type: string
                        example: "185"
                      market_value_in_eur:
                        type: integer
                        example: 1000000
                      highest_market_value_in_eur:
                        type: integer
                        example: 7500000
                      contract_expiration_date:
                        type: string
                        example: "2026-06-30 00:00:00"
                      agent_name:
                        type: string
                        nullable: true
                      image_url:
                        type: string
                        example: "https://img.a.transfermarkt.technology/portrait/header/419061-1665388396.png?lm=1"
                      url:
                        type: string
                        example: "https://www.transfermarkt.co.uk/vladyslav-supryaga/profil/spieler/419061"
                      __v:
                        type: integer
                        example: 0
                      stats:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            goals:
                              type: integer
                            assists:
                              type: integer
                            yellow_cards:
                              type: integer
                            red_cards:
                              type: integer
                            minutes_played:
                              type: integer
                            appearances:
                              type: integer
                            competitionName:
                              type: string
                      lineupsCount:
                        type: object
                        additionalProperties:
                          type: integer
                      totalStats:
                        type: object
                        properties:
                          goals:
                            type: integer
                          assists:
                            type: integer
                          yellow_cards:
                            type: integer
                          red_cards:
                            type: integer
                          minutes_played:
                            type: integer
                      market_values_in_eur:
                        type: array
                        items:
                          type: object
                          properties:
                            market_value_in_eur:
                              type: integer
                            date:
                              type: string
                      clubName:
                        type: string
                        example: "Futbolniy Klub Dynamo Kyiv"
    patch:
      tags:
        - Player
      description: Use to update one player
      parameters:
        - name: id
          in: path
          description: Player ID
          required: true
          example: 6574bc5d0a8453403804feea
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'
    delete:
      tags:
        - Player
      description: Use to delete one player
      parameters:
        - name: id
          in: path
          description: Player ID
          required: true
          example: 6574bc5d0a8453403804feea
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '/models/swaggerSchemas/player.yaml#/components/schemas/Player'
...
  /playerValuation:
    get:
      tags:
        - Player
        - PlayerValuation
      description: Use to request all player valuations
      parameters:
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/page'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/sort'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/limit'
        - $ref: '/models/swaggerSchemas/apiFeaturesParams.yaml#/parameters/fields'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/player_id'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/last_season'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/datetime'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/date'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/dateweek'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/market_value_in_eur'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/current_club_id'
        - $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/parameters/player_club_domestic_competition_id'
      responses:
        200:
          description: A successful response
          content:
              application/json:
                schema:
                    type: array
                    items:
                      $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/schemas/PlayerValuation'
    post:
      tags:
        - Player
        - PlayerValuation
      description: Use to create a new player valuation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/schemas/PlayerValuation'

      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '/models/swaggerSchemas/playerValuation.yaml#/components/schemas/PlayerValuation'
  /room:
    post:
      tags:
        - Room
      description: Use to create a new room
      requestBody:
        content:
          application/json:
            schema:
              $ref: '/models/swaggerSchemas/room.yaml#/components/schemas/ChatRoom'

      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '/models/swaggerSchemas/room.yaml#/components/schemas/ChatRoom'
...
  /room/search/{name}:
    get:
      tags:
        - Room
      description: Use to request all rooms that match the name param (the room's research is made checking if name param is a substring of the room's name)
      parameters:
        - name: name
          in: path
          description: Room name
          required: true
          example: room
      responses:
        200:
          description: A successful response
          content:
              application/json:
                schema:
                    type: array
                    items:
                      $ref: '/models/swaggerSchemas/room.yaml#/components/schemas/ChatRoom'
...
  /club/{id}:
    get:
      tags:
        - Club
      description: return some statistics for one club
      parameters:
        - name: id
          in: path
          description: club ID
          schema:
            type: integer
            example: 338
          required: true
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      clubId:
                        type: integer
                        example: 338
                      clubCode:
                        type: string
                        example: "dynamo-kiew"
                      name:
                        type: string
                        example: "Futbolniy Klub Dynamo Kyiv"
                      totalMarketValue:
                        type: integer
                        example: 88725000
                      squadSize:
                        type: string
                        example: "27"
                      averageAge:
                        type: string
                        example: "24.3"
                      foreignersNumber:
                        type: string
                        example: "3"
                      foreignersPercentage:
                        type: string
                        example: "11.1"
                      nationalTeamPlayers:
                        type: string
                        example: "8"
                      stadiumName:
                        type: string
                        example: "Valeriy Lobanovsky Stadion"
                      stadiumSeats:
                        type: string
                        example: "16873"
                      netTransferRecord:
                        type: string
                        example: "+€1.20m"
                      coachName:
                        type: string
                        nullable: true
                      lastSeason:
                        type: string
                        example: "2023"
                      url:
                        type: string
                        example: "https://www.transfermarkt.co.uk/dynamo-kiew/startseite/verein/338"
                      domesticCompetition:
                        type: object
                        properties:
                          competitionId:
                            type: string
                            example: "UKR1"
                          competitionCode:
                            type: string
                            example: "premier-liga"
                          name:
                            type: string
                            example: "premier-liga"
                          subType:
                            type: string
                            example: "first_tier"
                          type:
                            type: string
                            example: "domestic_league"
                          countryId:
                            type: string
                            example: "177"
                          countryName:
                            type: string
                            example: "Ukraine"
                          domesticLeagueCode:
                            type: string
                            example: "UKR1"
                          confederation:
                            type: string
                            example: "europa"
                          url:
                            type: string
                            example: "https://www.transfermarkt.co.uk/premier-liga/startseite/wettbewerb/UKR1"
                      players:
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                              example: "6574baf3975bb4cdf6a02818"
                            first_name:
                              type: string
                              nullable: true
                            last_name:
                              type: string
                              example: "Dudu"
                            country_of_citizenship:
                              type: string
                              example: "Brazil"
                            date_of_birth:
                              type: string
                              example: "1992-01-07"
                            position:
                              type: string
                              example: "Attack"
                            market_value_in_eur:
                              type: integer
                              example: 9000000
                            contract_expiration_date:
                              type: string
                              example: "2026-01-31 00:00:00"
                            id:
                              type: string
                              example: "6574baf3975bb4cdf6a02818"
  /search/{text}:
    get:
      tags:
        - Player
        - Club
      description: return players and clubs that match the text param (the research is made checking if text param is a substring of the player's full name or club's name)
      parameters:
        - name: text
          in: path
          description: text to search
          required: true
          example: messi
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  clubs:
                    type: array
                    items:
                      type: object
                      properties:
                        clubId:
                          type: integer
                          example: 12
                        name:
                          type: string
                          example: "Associazione Sportiva Roma"
                  players:
                    type: array
                    items:
                      type: object
                      properties:
                        playerId:
                          type: string
                          example: "6574baf3975bb4cdf6a00bde"
                        name:
                          type: string
                          example: "Thomas Hansen"
  /club:
    get:
      tags:
        - Club
      description: Use to request all clubs
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          example: 1
        - name: pagesize
          in: query
          description: Page size
          required: false
          example: 100

      responses:
        200:
          description: A successful response
          content:
              application/json:
                schema:
                    type: array
                    items:
                      $ref: '/models/swaggerSchemas/club.yaml#/Club'